pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.0"
python = "^3.8"
django-rest-framework = "^0.1.0"
djongo = "^1.3.6"
requests = "^2.26.0"
rabbitmq = "^0.2.0"
celery = "^5.2.1"
drf-writable-nested = "^0.6.3"
drf-yasg2 = "^1.19.4"
drf-yasg = "^1.20.0"
[tool.poetry.dev-dependencies]
